The second part of my change to hoist everything in the distcc/ directory up to the top level: delete the distcc directory (I already copied it all to the top level in my last commit).
This requires changing all files that refer to distcc/FOO to refer
directly to FOO instead.  Mostly this is Makefile.in and setup.py, but
there are a few other changes in scripts that have to hard-code paths,
like configure.ac and pump.in.  Also, several files changed their
comments to refer to the new location, which isn't strictly necessary
but I did anyway.

The only other change I made was to get rid of the hack I had put in
before, which renamed distcc to distcc-bin.  I had needed to do that
before, temporarily, because "distcc" was already the name of a
directory, and couldn't also be the name of a binary.  Now that the
distcc directory is going away, I can rename the target back to
distcc.

Tested by running './autogen; ./configure; make maintainer-check'.
I also ran 'make pump-check' and 'make distcheck'.

To be reviewed (post-facto) by klarlund@google.com and fergus@google.com