Previously, 'make check' was a noop because we do not require python to be installed for distcc, but the tests require python.  Therefore, to run the tests you had to run 'make maintainer-check' instead (since maintainers of distcc *do* have to have python installed).
Now, I have 'make check' check if python is installed.  If so, it runs
the tests.  Otherwise, it prints a message and exits.

Tested by running 'make check' and 'make installcheck' manually, on
systems with the various pythons installed as well as a few without.

Code reviewed by fergus@google.com