Fix what appears to be a bug in the handling of DISTCC_CMDLIST.
The documentation is not very clear, but distccd was documented
as matching against the last DISTCC_CMDLIST_NUMWORDS words of the
command, where I think by words they meant directory components.
For example, if the file refered to by DISTCC_CMDLIST contains the line
"/usr/bin/i686-blah-blah/gcc", and DISTCC_CMDLIST_NUMWORDS=2, only
commands which end in "/i686-blah-blah/gcc" would be matched.
The bug is that a command "i686-blah-blah/gcc" should be considered
to match.

Likewise, if DISTCC_CMDLIST contains the line "/usr/bin/gcc", and
DISTCC_CMDLIST_NUMWORDS is 1 (or not set), then plain "gcc" should be
allowed to match.

Reviewers: Dan Kegel, Craig Silverstein