Fix freak name collision on OSs that do not distinguish upper and lowercase in file names.
The symbol 'TESTING' is potentially unevaluated according to the
overapproximation semantics of the include server. So if this symbol is used in
a file name and if 'TESTING' really means 'testing' according to the OS and if
in fact 'testing' does exist as file, then there is trouble.

Revievewers: csilvers@google.com

Testing:
  make include-server-maintainer-check (on Linux)